sql30: Fix Python2 / 3 Column order issue.
Change :

In Python2 and Python 3.5 , dictionary traversal order is not same as
insertion order. It causes issues such as columns are created in certain
order but when updating or reading values, they come in different order.

Change here fixes this issue by determining and preserving the order of
creation of columns. This change doesn't impact Python 3.6 and onwards
but adds availability of sql30 on Python 2 and 3.5 as well.

If required, users can provide their own order of columns in table.