sql30: Allow sql30 users to chose DB files location based on env var.
sql30 users can already use "db_loc" parameter to specify the path where
DB dile shall be placed. In this change we allow users to specify an
environment variable, SQL30_DB_DIR, and with that all the db files shall
be written only in this location.

Note that it overwrites db_loc parameter as well.